+++ nexus-repository-helm-it/src/test/java/org/sonatype/nexus/plugins/helm/internal/HelmRepositoriesApiResourceTest.java
@@ -21,7 +21,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 
 public class HelmRepositoriesApiResourceTest
-    extends ResourceITSupport
+    extends HelmITSupport
 {
 
   @Configuration
